+++ react-ecommerce/src/App.js
@@ -27,6 +27,7 @@ import UserProfile from "./features/user/components/UserProfile";
 import UserProfilePage from "./pages/UserProfilePage";
 import { fetchLoggedInUserAsync } from "./features/user/userSlice";
 import Logout from "./features/auth/components/Logout";
+import ForgotPasswordPage from "./pages/ForgotPasswordPage";
 
 const router = createBrowserRouter([
   {
@@ -97,6 +98,12 @@ const router = createBrowserRouter([
     path: "logout",
     element: (
      <Logout />
+    ),
+  },
+  {
+    path: "forgot-password",
+    element: (
+     <ForgotPasswordPage />
     ),
   },
   {
