+++ react-ecommerce/src/pages/ForgotPasswordPage.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import ForgotPassword from '../features/auth/components/ForgotPassword'
+
+function ForgotPasswordPage() {
+  return (
+    <div>
+        <ForgotPassword />
+    </div>
+  )
+}
+
+export default ForgotPasswordPage
