+++ react-ecommerce/src/pages/LoginPage.js
@@ -1,140 +1,12 @@
 import React from "react";
-import { useForm } from "react-hook-form";
-import { Link, Navigate } from "react-router-dom";
-import {
-  checkloggedInUserAsync,
-  errorLoggedInUser,
-  selectedLoggedInUser,
-} from "../features/auth/authSlice";
-import { useDispatch, useSelector } from "react-redux";
+import Login from "../features/auth/components/Login";
 
 function LoginPage() {
-  const {
-    register,
-    handleSubmit,
-    watch,
-    formState: { errors },
-  } = useForm();
-
-  const user = useSelector(selectedLoggedInUser);
-  const loginError = useSelector(errorLoggedInUser);
-  const dispatch = useDispatch();
 
   return (
-    <>
-      {user && <Navigate to="/" replace={true} />}
-      <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8 bg-white">
-        <div className="sm:mx-auto sm:w-full sm:max-w-sm">
-          <img
-            alt="Your Company"
-            src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
-            className="mx-auto h-10 w-auto"
-          />
-          <h2 className="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">
-            Sign in to your account
-          </h2>
-        </div>
-
-        <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
-          <form
-          noValidate
-            onSubmit={handleSubmit((data) => {
-              dispatch(
-                checkloggedInUserAsync({
-                  email: data.email,
-                  password: data.password,
-                })
-              );
-            })}
-            className="space-y-6"
-          >
-            <div>
-              <label
-                htmlFor="email"
-                className="block text-sm font-medium leading-6 text-gray-900"
-              >
-                Email address
-              </label>
-              <div className="mt-2">
-                <input
-                  id="email"
-                  {...register("email", {
-                    required: "email is required",
-                    pattern: {
-                      value: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/,
-                      message: "Please enter a valid email address.",
-                    },
-                  })}
-                  type="email"
-                  className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                />
-                <p className="text-red-500">{errors?.email?.message}</p>
-              </div>
-            </div>
-
-            <div>
-              <div className="flex items-center justify-between">
-                <label
-                  htmlFor="password"
-                  className="block text-sm font-medium leading-6 text-gray-900"
-                >
-                  Password
-                </label>
-                <div className="text-sm">
-                  <a
-                    href="#"
-                    className="font-semibold text-indigo-600 hover:text-indigo-500"
-                  >
-                    Forgot password?
-                  </a>
-                </div>
-              </div>
-              <div className="mt-2">
-                <input
-                  id="password"
-                  {...register("password", {
-                    required: "password is required",
-                    pattern: {
-                      value:
-                        /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!]).{8,}$/,
-                      message:
-                        "Password must be at least 8 characters long, include an uppercase letter, a lowercase letter, a number, and a special character.",
-                    },
-                  })}
-                  type="password"
-                  className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                />
-                {errors?.password && (
-                  <p className="text-red-500">{errors?.password?.message}</p>
-                )}
-                {loginError && (
-                  <p className="text-red-500">{loginError?.message}</p>
-                )}
-              </div>
-            </div>
-
-            <div>
-              <button
-                type="submit"
-                className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-              >
-                Log in
-              </button>
-            </div>
-          </form>
-
-          <p className="mt-10 text-center text-sm text-gray-500">
-            Not a member?
-            <Link
-              to="/signup"
-              className="font-semibold leading-6 text-indigo-600 hover:text-indigo-500"
-            >
-              Create an account
-            </Link>
-          </p>
-        </div>
-      </div>
-    </>
+    <div>
+      <Login />
+    </div>
   );
 }
